Settings Menu + Scriptable objects
Settings menu it's basics is setup. The menu has a button to go back, and a title to show the player they are currently in the settings scene. Goal of the game is mentioned here. Also made scriptable objects for enemies and towers. Enemy their scriptable objects have been implemented in the game and now work. Added a wave state as well. Using the game state textbox, the game tells you can the wave has been defeated. (Have to test later how the game reacts after defeating the last wave).